+++ docs/release_notes.rst
@@ -1,6 +1,13 @@
 Release Notes
 =============
 
+v6.1.0
+------
+
+* Added `storage_utils.getsize` to public API (along with
+  `storage_utils.FileSystemPath.getsize` and
+  `storage_utils.swift.SwiftPath.getsize`). Port of `os.path.getsize`.
+
 v6.0.0
 ------